fix: 설비사진/현장사진 S3Key 발급 함수 분리
